+++ flexiapi/resources/views/parts/form/toggle.blade.php
@@ -1,5 +1,5 @@
 <div class="checkbox">
-    <input id="{{ $key }}" type="checkbox" @if ($object->$key || (isset($reversed) && $reversed && !$object->$key)) checked @endif name="{{ $key }}">
+    <input id="{{ $key }}" type="checkbox" @if ((!isset($reversed) && $object->$key) || (isset($reversed) && $reversed && !$object->$key)) checked @endif name="{{ $key }}">
     <label for="{{ $key }}"></label>
     <div>
         <p>{{ $label }}</p>