+++ CHANGELOG.md
@@ -21,6 +21,7 @@ v1.7
 - Fix FLEXIAPI-269 Update the IsNotPhoneNumber rule to use a better phone number validator
 - Fix FLEXIAPI-258 Move DotEnv instance configurations in the Spaces table
 - Fix FLEXIAPI-270 Call the static $apnsTypes attribute in FlexisipPusherConnector
+- Fix FLEXIAPI-271 Handle properly reversed attributes in objects
 
 v1.6
 ----